moved audio tokenizer code to the main file
